Scripts and session files in /tmp directory
Scripts from src/main/resources are extracted and installed as new files
in /tmp/tg-monitor/telethon/scripts. This is necessary, because when
this project is packed as a JAR file, and used as a dependency by
another code base, the scripts are not files anymore and can't just be
executed like when the project is used in unpacked form.

Session files are now saved under /tmp/tg-monitor/telethon/sessions.
Otherwise they would be saved in the current working directory, which is
wherever the command to start a program that uses this project is
launched from.

The /tmp directory seems to be cleared on system reboots on most sytems,
but this should be OK for the current case. Otherwise, there is also
/var/tmp which is supposed to persist across system reboots.

Also, the TelegramImpl class has been improved by extracting script and
session logic into their own classes, made available to the TelegramImpl
class by the ScriptManager and SessionManager interfaces, respectively.