Adapt to new way of reading messages
The monitor_peer.py script has been changed to read_messages.py, and
there is only one process executing this script running for the entire
application.

This process writes all the received messages of the logged-in account
to stdout. A thread in the Java application reads the output of this
process. It checks for each received message whether this message is
from a peer that is currently set in the Peers entity. If yes, then the
message is mapped to a TelegramMessage object and passed to the
Interactor. If no, then nothing is done with this message and the thread
keeps listening for then next received message output by the process.

This actually simplifies the implementation a lot. There is only one
process, one thread, and one session file. The only thing you could say
against it is that this process and thread may possibly come under very
heavy load, if the Telegram account has an extremely high message
arrival rate. But I don't think that this will be a real problem in
normal practical scenarios.